Day9 IntMode preperations
This moves IntCode to work on 64bit ints as well as
introduces relative address mode.

+++ advent_of_code/2019/day5/day5.cc
@@ -11,8 +11,8 @@ absl::StatusOr<std::vector<std::string>> Day5_2019::Part1(
   absl::StatusOr<IntCode> codes = IntCode::Parse(input_text);
   if (!codes.ok()) return codes.status();
 
-  std::vector<int> input = {1};
-  std::vector<int> output;
+  std::vector<int64_t> input = {1};
+  std::vector<int64_t> output;
   if (absl::Status st = codes->Run(input, &output); !st.ok()) return st;
 
   for (int i = 0; i < output.size() - 1; ++i) {
@@ -29,8 +29,8 @@ absl::StatusOr<std::vector<std::string>> Day5_2019::Part2(
   absl::StatusOr<IntCode> codes = IntCode::Parse(input_text);
   if (!codes.ok()) return codes.status();
 
-  std::vector<int> input = {5};
-  std::vector<int> output;
+  std::vector<int64_t> input = {5};
+  std::vector<int64_t> output;
   if (absl::Status st = codes->Run(input, &output); !st.ok()) return st;
 
   for (int i = 0; i < output.size() - 1; ++i) {
